render: move the usage-line helpers into their own module
detail_pane/mod.rs reached 801 code lines before its test module, one over the
ceiling. The six usage-line functions move to detail_pane/usage_form.rs.
No behavior change.
